updated test case

+++ pluto-plugins/plugins/network/lib/src/test/java/com/pluto/plugins/network/internal/bandwidth/core/BandwidthLimitSocketFactoryTest.kt
@@ -60,7 +60,8 @@ class BandwidthLimitSocketFactoryTest {
 
         val uploadTimeSeconds = (endTime - startTime) / 1000.0
         Assert.assertEquals(
-            dataSizeBytes / ThrottledOutputStream.maxBytesPerSecond, uploadTimeSeconds.toInt()
+            (dataSizeBytes / ThrottledOutputStream.maxBytesPerSecond).toInt(),
+            uploadTimeSeconds.toInt()
         )
     }
